payment amount datatype changed to float

+++ JEDI-2.0-FlipFit-GRP-I-JAVA-POS/src/com/flipkart/bean/Payment.java
@@ -4,7 +4,7 @@ public class Payment {
 	
 	private String paymentid;
 	private String bookingid;
-	private int amountPaid;
+	private float amountPaid;
 	
 	public String getPaymentid() {
 		return paymentid;
@@ -18,7 +18,7 @@ public class Payment {
 	public void setBookingid(String bookingid) {
 		this.bookingid = bookingid;
 	}
-	public int getAmountPaid() {
+	public float getAmountPaid() {
 		return amountPaid;
 	}
 	public void setAmountPaid(int amountPaid) {
